+++ TokenStepSwift/Tests/TokenStepSwiftTests/UpdateServiceVersionTests.swift
@@ -9,7 +9,8 @@ final class UpdateServiceVersionTests: XCTestCase {
     func testPrereleaseIdentifiersFollowSemanticVersionOrder() {
         XCTAssertLessThan(Version("0.1.0-beta.1"), Version("0.1.0-beta.2"))
         XCTAssertLessThan(Version("0.1.0-beta.2"), Version("0.1.0-beta.3"))
-        XCTAssertLessThan(Version("0.1.0-beta.3"), Version("0.1.0-rc.1"))
+        XCTAssertLessThan(Version("0.1.0-beta.3"), Version("0.1.0-beta.4"))
+        XCTAssertLessThan(Version("0.1.0-beta.4"), Version("0.1.0-rc.1"))
     }
 
     func testBuildMetadataDoesNotChangePrecedence() {